Walk back approach: Keep config entities, but use core API to suppress.
The approach with content entities got into issue due to varying
identifiers in the derived `migrate_plus` entities that would have to be
updated, as rederiving from the migrations-in-code could drift from what
the old `migrate_plus` entities described.

Given the issue is really that we want to prevent these entities from
hitting config imports/exports, and we needed a solution for it for the
`migrate_plus` entities in any case, it seems easiest just to keep the
request entities as config entities for now. If/when we rework away from
`migrate_plus` and `migrate_tools` for GUI, and have some lifecycle
expectations baked in around request entities, it might make sense to
revisit.