remove rollback feature
initially added in https://github.com/discoverygarden/islandora_spreadsheet_ingest/pull/109

found bugs with it deleting taxonomy terms still referenced by existing
objects.

Steps to reproduce and scope of concern:

    Run a spreadsheet ingest referencing or creating terms in taxonomies that allow for terms to be created on ingest.

    Run a second spreadsheet ingest that creates more objects referencing the same terms as are referenced/created previously.

    For either of the two ingests, use rollback feature in the UI that was added in https://github.com/discoverygarden/islandora_spreadsheet_ingest/pull/109

    The rollback will delete taxonomy terms still being referenced by existing content, but only for taxonomies that allow for terms to be created on ingest.

Note about what I mean by taxonomies that allow for terms to be created on ingest:

This does not appear to affect taxonomies that are more locked down and only allow ingests to reference existing terms. For example, taxonomies like ‘Islandora Models’ or ‘Rights Statements’ need to already have the terms being referenced on ingest in order for the ingest to go through, and the rollback does not delete terms from those taxonomies, only from taxonomies that can have terms created by object ingests (regardless of if the term was created on ingest or not).

Confirmed previously existing rollback functionality is not a concern, so drush command is fine and individual migration rollbacks from migration UI is fine, just the new spreadsheet ingest rollback as noted above is the problem.